// models/create-call-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bandwidth
 * Bandwidth\'s Communication APIs
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: letstalk@bandwidth.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CallbackMethodEnum } from './callback-method-enum';

export interface CreateCallResponse {
    /**
     * The id of the application associated with the `from` number.
     */
    'applicationId': string;
    /**
     * The bandwidth account ID associated with the call.
     */
    'accountId': string;
    /**
     * Programmable Voice API Call ID.
     */
    'callId': string;
    /**
     * Recipient of the outgoing call.
     */
    'to': string;
    /**
     * Phone number that created the outbound call.
     */
    'from': string;
    /**
     * The time at which the call was accepted into the queue.
     */
    'enqueuedTime'?: string | null;
    /**
     * The URL to update this call\'s state.
     */
    'callUrl': string;
    /**
     * The timeout (in seconds) for the callee to answer the call after it starts ringing.
     */
    'callTimeout'?: number;
    /**
     * This is the timeout (in seconds) to use when delivering webhooks for the call.
     */
    'callbackTimeout'?: number;
    /**
     * Custom tag value.
     */
    'tag'?: string | null;
    'answerMethod': CallbackMethodEnum | null;
    /**
     * URL to deliver the `answer` event webhook.
     */
    'answerUrl': string;
    'answerFallbackMethod'?: CallbackMethodEnum | null;
    /**
     * Fallback URL to deliver the `answer` event webhook.
     */
    'answerFallbackUrl'?: string | null;
    'disconnectMethod': CallbackMethodEnum | null;
    /**
     * URL to deliver the `disconnect` event webhook.
     */
    'disconnectUrl'?: string | null;
    /**
     * Basic auth username.
     */
    'username'?: string | null;
    /**
     * Basic auth password.
     */
    'password'?: string | null;
    /**
     * Basic auth username.
     */
    'fallbackUsername'?: string | null;
    /**
     * Basic auth password.
     */
    'fallbackPassword'?: string | null;
    /**
     * The priority of this call over other calls from your account.
     */
    'priority'?: number | null;
}